requirements.txt
# Core dependencies for lolo-ai-documents
# Generated from pyproject.toml

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI/LLM
langchain>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.20

# Database (MySQL only - sessions stored in JUDICIAL_AI_DOCUMENT_SESSION table)
aiomysql>=0.2.0

# AWS
boto3>=1.34.0
aioboto3>=12.0.0

# Document generation
python-docx>=1.1.0

# Document extraction (PDF, Word, Excel, Images)
pypdf>=4.0.0
pymupdf>=1.24.0
openpyxl>=3.1.0
pillow>=10.0.0
pillow-heif>=0.15.0  # HEIC/HEIF image support (iPhone photos)

# Anthropic SDK (for Claude Vision API - image analysis)
anthropic>=0.39.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
loguru>=0.7.0
python-dotenv>=1.0.0

# Rate limiting
slowapi>=0.1.9

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Development
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
